my-app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
psycopg2-binary
python-dotenv
